Add files via upload
This is a short script that may be useful for anyone needing to check the minor allele frequency (MAF) of a variant in their cohort.
It takes as input .map and .ped Plink next-generation sequencing files. These should first be processed through quality control stages.
The script can easily be modified to analyse differences in the MAF between different groups (disease, ethnicity, etc.).